refactor(mobile): address You-page review
- Remove no-op identity `select` from usePublicProfile.
- Drop dead `onClose` no-op props from YouFilterSheet + FollowListSheet
  (both sheets close via their own ref; no close cleanup needed). CommentSheet
  and LogbookEditSheet keep onClose since they reset state.
- useYouProfileData: replace inaccessible fromDate/toDate state with empty
  constants + TODO (custom date-range UI is a follow-up); stop exposing the
  setters.
- Move BAR_CONTENT_HEIGHT/TAB_BAR_HEIGHT into theme/layout.ts as the source of
  truth; persistent-queue-bar + BlurTabBar re-export for back-compat, and the
  You tabs now import layout metrics from theme/layout instead of queue-control
  internals.
- CLAUDE.md: list @boardsesh/board-react and @boardsesh/profile-stats in the
  monorepo structure.
